[Integration Test] Fixing CR comments (test settings file, packaging etc.)
