--- web/src/integTest/java/com/microsoft/applicationinsights/framework/telemetries/PerformanceCounterTelemetryItem.java
@@ -1,66 +1,0 @@
-/*
- * ApplicationInsights-Java
- * Copyright (c) Microsoft Corporation
- * All rights reserved.
- *
- * MIT License
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the ""Software""), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify, merge,
- * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
- * persons to whom the Software is furnished to do so, subject to the following conditions:
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
- * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
- */
-
-package com.microsoft.applicationinsights.framework.telemetries;
-
-import org.json.JSONException;
-import org.json.JSONObject;
-
-/**
- * Created by moralt on 05/05/2015.
- */
-public class PerformanceCounterTelemetryItem extends TelemetryItem {
-    private static final String[] propertiesToCompare = new String[] {
-            "category",
-            "instance",
-    };
-
-    public PerformanceCounterTelemetryItem() {
-        super(DocumentType.PerformanceCounters);
-    }
-
-    public PerformanceCounterTelemetryItem(JSONObject json) throws JSONException {
-        this();
-
-        initPerformanceCounterTelemetryItem(json);
-    }
-
-    protected String[] getDefaultPropertiesToCompare() {
-        return propertiesToCompare;
-    }
-
-    /**
-     * Converts JSON object to PerformanceCounter TelemetryItem
-     * @param json The JSON object
-     * @return A TelemetryItem
-     */
-    private void initPerformanceCounterTelemetryItem(JSONObject json) throws JSONException {
-        System.out.println("Converting JSON object to telemetry item PerformanceCounterTelemetryItem");
-
-        JSONObject performanceCounterProperties = json.getJSONArray("performanceCounter").getJSONObject(0);
-
-        String category = performanceCounterProperties.getString("categoryName");
-        String instance = performanceCounterProperties.getString("instanceName");
-
-        this.setProperty("category", category);
-        this.setProperty("instance", instance);
-    }
-}
