+++ web/src/integTest/java/com/microsoft/applicationinsights/test/framework/TelemetryBlob.java
@@ -1,0 +1,45 @@
+/*
+ * ApplicationInsights-Java
+ * Copyright (c) Microsoft Corporation
+ * All rights reserved.
+ *
+ * MIT License
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the ""Software""), to deal in the Software
+ * without restriction, including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+ * persons to whom the Software is furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+ * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
+ * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ */
+
+package com.microsoft.applicationinsights.test.framework;
+
+import com.microsoft.applicationinsights.test.framework.telemetries.DocumentType;
+
+/**
+ * Created by yonisha on 6/21/2015.
+ */
+public class TelemetryBlob {
+    private final DocumentType docType;
+    private final String blobUri;
+
+    public TelemetryBlob(DocumentType docType, String blobUri) {
+        this.docType = docType;
+        this.blobUri = blobUri;
+    }
+
+    public DocumentType getDocType() {
+        return this.docType;
+    }
+
+    public String getBlobUri() {
+        return this.blobUri;
+    }
+}
