+++ web/src/integTest/java/com/microsoft/applicationinsights/test/framework/TestEnvironment.java
@@ -1,4 +1,6 @@
-package com.microsoft.applicationinsights.framework;
+package com.microsoft.applicationinsights.test.framework;
+
+import com.microsoft.applicationinsights.test.framework.utils.PropertiesUtils;
 
 import java.io.*;
 import java.util.Properties;
@@ -23,7 +25,7 @@ public class TestEnvironment {
             throw new RuntimeException("Environment variable " + configurationEnvironmentVariable + " not defined");
         }
 
-        envProps = loadPropertiesFromFile(envPropsFilePath);
+        envProps = PropertiesUtils.loadPropertiesFromFile(envPropsFilePath);
     }
 
     public String getApplicationServer() {
@@ -44,34 +46,5 @@ public class TestEnvironment {
 
     public String getApplicationStorageExportQueueName() {
         return envProps.getProperty(KEY_APPLICATION_STORAGE_EXPORT_QUEUE_NAME);
-    }
-
-    public static Properties loadPropertiesFromFile(String filePath) throws IOException
-    {
-        File configFile = new File(filePath);
-        if (!configFile.exists()) {
-            throw new FileNotFoundException("Configuration file \"" + filePath + "\" could not be found");
-        }
-
-        System.out.println("Found configuration file at: " + filePath);
-
-        InputStream inputStream = new FileInputStream(filePath);
-
-        return readFromStream(inputStream);
-    }
-
-    private static Properties readFromStream(InputStream inputStream) throws IOException {
-        Properties props = new Properties();
-
-        if (inputStream != null)
-        {
-            try {
-                props.load(inputStream);
-            } finally {
-                inputStream.close();
-            }
-        }
-
-        return props;
     }
 }
