+++ web/src/integTest/java/com/microsoft/applicationinsights/test/framework/telemetries/EventTelemetryItem.java
@@ -19,38 +19,27 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.framework;
+package com.microsoft.applicationinsights.test.framework.telemetries;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+import java.net.URISyntaxException;
 
 /**
- * Created by yonisha on 6/16/2015.
+ * Created by yonisha on 6/21/2015.
  */
-public class UriWithExpectedResult {
+public class EventTelemetryItem extends TelemetryItem {
 
-    private final int expectedResponseCode;
-    private final String uri;
-    private final String expectedRequestName;
-    private final String runId;
-
-    public UriWithExpectedResult(String uri, String runId, int expectedResponseCode, String expectedRequestName) {
-        this.uri = uri;
-        this.runId = runId;
-        this.expectedResponseCode = expectedResponseCode;
-        this.expectedRequestName = expectedRequestName;
+    public EventTelemetryItem() {
+        super(DocumentType.Event);
     }
 
-    public String getUri() {
-        return this.uri;
+    public EventTelemetryItem(JSONObject json) throws URISyntaxException, JSONException {
+        super(DocumentType.Event, json);
     }
 
-    public String getRunId() {
-        return this.runId;
-    }
-
-    public int getExpectedResponseCode() {
-        return this.expectedResponseCode;
-    }
-
-    public String getExpectedRequestName() {
-        return this.expectedRequestName;
+    @Override
+    protected String[] getDefaultPropertiesToCompare() {
+        return new String[0];
     }
 }
