+++ web/src/integTest/java/com/microsoft/applicationinsights/test/framework/utils/PropertiesUtils.java
@@ -1,0 +1,65 @@
+/*
+ * ApplicationInsights-Java
+ * Copyright (c) Microsoft Corporation
+ * All rights reserved.
+ *
+ * MIT License
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the ""Software""), to deal in the Software
+ * without restriction, including without limitation the rights to use, copy, modify, merge,
+ * publish, distribute, sublicense, and/or sell copies of the Software, and to permit
+ * persons to whom the Software is furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+ * PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE
+ * FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+ * DEALINGS IN THE SOFTWARE.
+ */
+
+package com.microsoft.applicationinsights.test.framework.utils;
+
+import java.io.*;
+import java.util.Properties;
+
+/**
+ * Created by amnonsh on 5/28/2015.
+ */
+public class PropertiesUtils {
+
+    public static Properties loadPropertiesFromResource(String resourceName) throws IOException
+    {
+        ClassLoader classLoader = PropertiesUtils.class.getClassLoader();
+        InputStream inputStream = classLoader.getResourceAsStream(resourceName);
+        return loadProperties(inputStream);
+    }
+
+    public static Properties loadPropertiesFromFile(String filePath) throws IOException
+    {
+        File configFile = new File(filePath);
+        if (!configFile.exists()) {
+            throw new FileNotFoundException("Configuration file \"" + filePath + "\" could not be found");
+        }
+        System.out.println("Configuration file located in " + filePath);
+
+        InputStream inputStream = new FileInputStream(filePath);
+        return loadProperties(inputStream);
+    }
+
+    private static Properties loadProperties(InputStream inputStream) throws IOException {
+        Properties props = new Properties();
+
+        if (inputStream != null)
+        {
+            try {
+                props.load(inputStream);
+            } finally {
+                inputStream.close();
+            }
+        }
+
+        return props;
+    }
+}
