+++ web/src/integTest/resources/testSettings.properties
@@ -1,0 +1,3 @@
+maxWaitTime=180
+keyPollingInterval=3
+keyMessageBatchSize=32
